Moving array versions of track check related functions
