+++ tests/test_track_check_qc_functions.py
@@ -19,18 +19,16 @@ from marine_qc.qc_sequential_reports import (
     calculate_course_parameters,
     calculate_speed_course_distance_time_difference,
     calculate_speed_course_distance_time_difference_array,
-    lat_lon_from_course_and_distance_array,
-    course_between_points_array,
     forward_discrepancy_array,
     backward_discrepancy_array,
     forward_discrepancy,
     calculate_midpoint,
     calculate_midpoint_array,
-    time_differences_array,
-    sphere_distance_array,
-    intermediate_point_array,
     do_track_check_array,
 )
+from marine_qc.spherical_geometry import sphere_distance_array, intermediate_point_array, course_between_points_array, \
+    lat_lon_from_course_and_distance_array
+from marine_qc.time_control import time_differences_array
 
 
 def generic_frame(in_pt):
